Add common recommender class
Refactors MostPopular
